+++ src/main/java/com/busyqa/crm/repo/AcademicsRepositoryI.java
@@ -10,7 +10,7 @@ public interface AcademicsRepositoryI {
     // COURSE
     void addCourse(Course course);
     List<Course> getAllCourse();
-    Course getCourseById(int id);
+    Course getCourseById(long id);
     boolean courseExist(String name);
     Optional<Course> findByCourseName(String name);
 
@@ -18,24 +18,24 @@ public interface AcademicsRepositoryI {
     // CLASS
     void addClassByCourseId(Integer classId, Integer courseId);
     List<Class> getAllClass();
-    Class getClassById(int id);
+    Class getClassById(long id);
 
     // COURSE SCHEDULE
     List<CourseSchedule> getAllCourseSchedule();
     void addCourseSchedule(CourseSchedule courseSchedule);
-    CourseSchedule getCourseScheduleById(int id);
+    CourseSchedule getCourseScheduleById(long id);
     boolean courseScheduleExist(String name);
 
     // TRAINER
     List<Trainer> getAllTrainer();
     void addTrainer(Trainer trainer);
-    Trainer getTrainerById(int id);
+    Trainer getTrainerById(long id);
     boolean trainerExist(String name);
 
     // TRAINING LOCATION
     List<TrainingLocation> getAllTrainingLocation();
     void addTrainingLocation(TrainingLocation trainingLocation);
-    TrainingLocation getTrainingLocationById(int id);
+    TrainingLocation getTrainingLocationById(long id);
     boolean trainingLocationExist(String name);
